refactor(shared): move type aliases from constants to interfaces
Relocate the six shared type aliases that were declared in constants
files into the matching .interface.ts, per the placement convention
documented in LFXV2-2570:

- CountryCode      -> interfaces/countries.interface.ts (new)
- USState          -> interfaces/states.interface.ts (new)
- TShirtSize       -> interfaces/tshirt-sizes.interface.ts (new)
- AllowedLogoMimeType -> interfaces/crowdfunding.interface.ts
- CombinedSurveyStatus -> interfaces/survey.interface.ts
- ProductType      -> interfaces/training.interface.ts

Derived aliases import their source constant type-only (erased at
compile time, no runtime cycle). Constants files that still reference a
moved alias re-import it type-only. Import sites and the interfaces
barrel are updated accordingly. No behavior change.

Refs: LFXV2-2573
