+++ packages/shared/src/constants/training.constants.ts
@@ -80,7 +80,6 @@ export const DESCENDING_DEFAULT_ORG_TRAINING_SORT_FIELDS: ReadonlySet<string> = 
 
 export const TRAINING_PRODUCT_TYPE = 'Training' as const;
 export const CERTIFICATION_PRODUCT_TYPE = 'Certification' as const;
-export type ProductType = typeof TRAINING_PRODUCT_TYPE | typeof CERTIFICATION_PRODUCT_TYPE;
 
 export const CONTINUE_LEARNING_URL = 'https://trainingportal.linuxfoundation.org/learn/dashboard';
 export const COURSE_URL_PREFIX = 'https://trainingportal.linuxfoundation.org/learn/course/';
